feat: 홈 서치 뷰 작업 중
- Search VC, View 생성 및 작업
