+++ Brandingdong/Brandingdong/Controller/Brandingdong/Controller/Home/Search/SearchViewController.swift
@@ -1,0 +1,31 @@
+//
+//  SearchViewController.swift
+//  Brandingdong
+//
+//  Created by 이진욱 on 2020/08/24.
+//  Copyright © 2020 jwlee. All rights reserved.
+//
+
+import UIKit
+
+class SearchViewController: UIViewController {
+  // MARK: - Property
+  private let searchController = UISearchController(searchResultsController: nil)
+  
+  // MARK: - LifeCycle
+  
+  override func viewDidLoad() {
+    super.viewDidLoad()
+  }
+  
+  // MARK: - Setup Layout
+  
+  private func setUI() {
+    
+  }
+  
+  private func setConstraints() {
+    
+  }
+
+}
